backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
google-auth==2.35.0
google-api-python-client==2.149.0
